+++ examples/Card/Controller.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * This is the return controller for the Card example.
- * It is called when the client is redirected back to the shop from the 3ds page.
+ * This is the controller for the Card example.
+ * It is called when the pay button on the index page is clicked.
  *
- * Copyright (C) 2018 heidelpay GmbH
+ * Copyright (C) 2019 heidelpay GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,15 +27,17 @@
 /** Require the constants of this example */
 require_once __DIR__ . '/Constants.php';
 
+/** @noinspection PhpIncludeInspection */
 /** Require the composer autoloader file */
 require_once __DIR__ . '/../../../../autoload.php';
 
 use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
-use heidelpayPHP\Resources\TransactionTypes\Authorization;
+use heidelpayPHP\Resources\Customer;
 
 session_start();
+session_unset();
 
 $clientMessage = 'Something went wrong. Please try again later.';
 $merchantMessage = 'Something went wrong. Please try again later.';
@@ -48,11 +50,15 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
     die();
 }
 
-// Retrieve the paymentId you remembered within the Controller
-if (!isset($_SESSION['PaymentId'])) {
-    redirect(FAILURE_URL, 'The payment id is missing.', $clientMessage);
+// You will need the id of the payment type created in the frontend (index.php)
+if (!isset($_POST['resourceId'])) {
+    redirect(FAILURE_URL, 'Resource id is missing!', $clientMessage);
 }
-$paymentId = $_SESSION['PaymentId'];
+$paymentTypeId   = $_POST['resourceId'];
+
+// These lines are just for this example
+$transactionType = $_POST['transaction_type'] ?? 'authorize';
+$use3Ds          = isset($_POST['3dsecure']) && ($_POST['3dsecure'] === '1');
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
@@ -60,23 +66,30 @@ try {
     $heidelpay = new Heidelpay(HEIDELPAY_PHP_PAYMENT_API_PRIVATE_KEY);
     $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    // Redirect to success if the payment has been successfully completed or is still in handled.
-    $payment = $heidelpay->fetchPayment($paymentId);
-    if (
-        $payment->isCompleted()         // <<---- in case of charge
-        || $payment->isPending()        // <<---- in case of authorize
-    ) {
+    // Create a charge/authorize transaction
+    // The 3D secured flag can be used to switch between 3ds and non-3ds.
+    // If your merchant is only configured for one of those you can omit the flag.
+    $customer     = new Customer('Linda', 'Heideich');
+    if ($transactionType === 'charge') {
+        $transaction = $heidelpay->charge(12.99, 'EUR', $paymentTypeId, RETURN_CONTROLLER_URL, $customer, null, null, null, $use3Ds);
+    } else {
+        $transaction = $heidelpay->authorize(12.99, 'EUR', $paymentTypeId, RETURN_CONTROLLER_URL, $customer, null, null, null, $use3Ds);
+    }
+
+    // You'll need to remember the paymentId for later in the ReturnController (in case of 3ds)
+    $_SESSION['PaymentId'] = $transaction->getPaymentId();
+    $_SESSION['ShortId'] = $transaction->getShortId();
+
+    // Redirect to the 3ds page or to success depending on the state of the transaction
+    $payment = $transaction->getPayment();
+    if (empty($transaction->getRedirectUrl()) && $transaction->isSuccess()) {
         redirect(SUCCESS_URL);
+    } elseif (!empty($transaction->getRedirectUrl()) && $transaction->isPending()) {
+        redirect($transaction->getRedirectUrl());
     }
 
     // Check the result message of the transaction to find out what went wrong.
-    $transaction = $payment->getAuthorization();
-    if ($transaction instanceof Authorization) {
-        $merchantMessage = $transaction->getMessage()->getCustomer();
-    } else {
-        $transaction = $payment->getChargeByIndex(0);
-        $merchantMessage = $transaction->getMessage()->getCustomer();
-    }
+    $merchantMessage = $transaction->getMessage()->getCustomer();
 } catch (HeidelpayApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
     $clientMessage = $e->getClientMessage();