+++ examples/IDeal/ReturnController.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * This is the controller for the Card example.
- * It is called when the pay button on the index page is clicked.
+ * This is the return controller for the iDeal example.
+ * It is called when the client is redirected back to the shop from the external page.
  *
  * Copyright (C) 2019 heidelpay GmbH
  *
@@ -27,16 +27,13 @@
 /** Require the constants of this example */
 require_once __DIR__ . '/Constants.php';
 
+/** @noinspection PhpIncludeInspection */
 /** Require the composer autoloader file */
 require_once __DIR__ . '/../../../../autoload.php';
 
 use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
-use heidelpayPHP\Resources\Customer;
-
-session_start();
-session_unset();
 
 $clientMessage = 'Something went wrong. Please try again later.';
 $merchantMessage = 'Something went wrong. Please try again later.';
@@ -49,15 +46,13 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
     die();
 }
 
-// You will need the id of the payment type created in the frontend (index.php)
-if (!isset($_POST['resourceId'])) {
-    redirect(FAILURE_URL, 'Resource id is missing!', $clientMessage);
-}
-$paymentTypeId   = $_POST['resourceId'];
+session_start();
 
-// These lines are just for this example
-$transactionType = $_POST['transaction_type'] ?? 'authorize';
-$use3Ds          = isset($_POST['3dsecure']) && ($_POST['3dsecure'] === '1');
+// Retrieve the paymentId you remembered within the Controller
+if (!isset($_SESSION['PaymentId'])) {
+    redirect(FAILURE_URL, 'The payment id is missing.', $clientMessage);
+}
+$paymentId = $_SESSION['PaymentId'];
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
@@ -65,30 +60,15 @@ try {
     $heidelpay = new Heidelpay(HEIDELPAY_PHP_PAYMENT_API_PRIVATE_KEY);
     $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    // Create a charge/authorize transaction
-    // The 3D secured flag can be used to switch between 3ds and non-3ds.
-    // If your merchant is only configured for one of those you can omit the flag.
-    $customer     = new Customer('Linda', 'Heideich');
-    if ($transactionType === 'charge') {
-        $transaction = $heidelpay->charge(12.99, 'EUR', $paymentTypeId, RETURN_CONTROLLER_URL, $customer, null, null, null, $use3Ds);
-    } else {
-        $transaction = $heidelpay->authorize(12.99, 'EUR', $paymentTypeId, RETURN_CONTROLLER_URL, $customer, null, null, null, $use3Ds);
-    }
-
-    // You'll need to remember the paymentId for later in the ReturnController (in case of 3ds)
-    $_SESSION['PaymentId'] = $transaction->getPaymentId();
-    $_SESSION['ShortId'] = $transaction->getShortId();
-
-    // Redirect to the 3ds page or to success depending on the state of the transaction
-    $payment = $transaction->getPayment();
-    if (empty($transaction->getRedirectUrl()) && $transaction->isSuccess()) {
+    // Redirect to success if the payment has been successfully completed.
+    $payment   = $heidelpay->fetchPayment($paymentId);
+    if ($payment->isCompleted()) {
         redirect(SUCCESS_URL);
-    } elseif (!empty($transaction->getRedirectUrl()) && $transaction->isPending()) {
-        redirect($transaction->getRedirectUrl());
     }
 
-    // Check the result message of the transaction to find out what went wrong.
-    $merchantMessage = $transaction->getMessage()->getCustomer();
+    // Check the result message of the charge to find out what went wrong.
+    $charge = $payment->getChargeByIndex(0);
+    $merchantMessage = $charge->getMessage()->getCustomer();
 } catch (HeidelpayApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
     $clientMessage = $e->getClientMessage();
@@ -96,3 +76,4 @@ try {
     $merchantMessage = $e->getMessage();
 }
 redirect(FAILURE_URL, $merchantMessage, $clientMessage);
+