+++ examples/index.php
@@ -59,12 +59,23 @@ require_once __DIR__ . '/../../../autoload.php';
             <div class="ui four cards">
                 <div class="card olive">
                     <div class="content">
-                        <div class="header">Credit Card</div>
+                        <div class="header">Card</div>
                         <div class="description">
                             You can try authorize and charge transaction with or without 3Ds.
                         </div>
                     </div>
-                    <div class="ui bottom attached green button" onclick="location.href='CreditCard/';">
+                    <div class="ui bottom attached green button" onclick="location.href='Card/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">Card Recurring</div>
+                        <div class="description">
+                            You can set a Card type to recurring in order to register it and charge later as well as implement recurring payments.
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='CardRecurring/';">
                         Try
                     </div>
                 </div>
@@ -77,6 +88,18 @@ require_once __DIR__ . '/../../../autoload.php';
                         </div>
                     </div>
                     <div class="ui bottom attached green button" onclick="location.href='EPSCharge/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">
+                            iDeal
+                        </div>
+                        <div class="description">
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='IDeal/';">
                         Try
                     </div>
                 </div>
@@ -142,6 +165,31 @@ require_once __DIR__ . '/../../../autoload.php';
                         </div>
                     </div>
                     <div class="ui bottom attached green button" onclick="location.href='PayPal/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">
+                            PayPal Recurring
+                        </div>
+                        <div class="description">
+                            You can set a Pay Pal type to recurring in order to register it and charge later as well as implement recurring payments.
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='PayPalRecurring/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">
+                            Sofort
+                        </div>
+                        <div class="description">
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='Sofort/';">
                         Try
                     </div>
                 </div>