+++ src/Heidelpay.php
@@ -334,6 +334,26 @@ class Heidelpay implements HeidelpayParentInterface
         return $this->resourceService->fetch($resource);
     }
 
+    //<editor-fold desc="Recurring">
+
+    /**
+     * Activate recurring payment for the given payment type (if possible).
+     *
+     * @param string|BasePaymentType $paymentType The payment to activate recurring payment for.
+     * @param string                 $returnUrl   The URL to which the customer gets redirected in case of a 3ds transaction
+     *
+     * @return mixed
+     *
+     * @throws HeidelpayApiException
+     * @throws RuntimeException
+     */
+    public function activateRecurringPayment($paymentType, $returnUrl)
+    {
+        return $this->resourceService->createRecurring($paymentType, $returnUrl);
+    }
+
+    //</editor-fold>
+
     //<editor-fold desc="Payment resource">
 
     /**
@@ -852,14 +872,16 @@ class Heidelpay implements HeidelpayParentInterface
     /**
      * Fetches a resource object based on the given event data.
      *
+     * @param string|null $eventJson
+     *
      * @return AbstractHeidelpayResource
      *
      * @throws HeidelpayApiException
      * @throws RuntimeException
      */
-    public function fetchResourceFromEvent(): AbstractHeidelpayResource
+    public function fetchResourceFromEvent($eventJson = null): AbstractHeidelpayResource
     {
-        return $this->webhookService->fetchResourceByWebhookEvent();
+        return $this->webhookService->fetchResourceByWebhookEvent($eventJson);
     }
 
     //</editor-fold>