+++ src/Resources/EmbeddedResources/BasketItem.php
@@ -58,6 +58,12 @@ class BasketItem extends AbstractHeidelpayResource
     /** @var string $title */
     protected $title = '';
 
+    /** @var string|null $subTitle */
+    protected $subTitle;
+
+    /** @var string|null $imageUrl */
+    protected $imageUrl;
+
     /**
      * BasketItem constructor.
      *
@@ -267,6 +273,43 @@ class BasketItem extends AbstractHeidelpayResource
     public function setTitle(string $title): BasketItem
     {
         $this->title = $title;
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getImageUrl()
+    {
+        return $this->imageUrl;
+    }
+
+    /**
+     * @param string|null $imageUrl
+     *
+     * @return BasketItem
+     */
+    public function setImageUrl($imageUrl): BasketItem
+    {
+        $this->imageUrl = $imageUrl;
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getSubTitle()
+    {
+        return $this->subTitle;
+    }
+
+    /**
+     * @param string|null $subTitle
+     * @return BasketItem
+     */
+    public function setSubTitle($subTitle): BasketItem
+    {
+        $this->subTitle = $subTitle;
         return $this;
     }