+++ test/integration/DummyResource.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * This service provides for functionalities concerning the mgw environment.
+ * This class defines a dummy resource used for unit tests.
  *
- * Copyright (C) 2019 heidelpay GmbH
+ * Copyright (C) 2018 heidelpay GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,19 +20,13 @@
  *
  * @author  Simon Gabriel <development@heidelpay.com>
  *
- * @package  heidelpayPHP/services
+ * @package  heidelpayPHP/test/unit
  */
-namespace heidelpayPHP\Services;
+namespace heidelpayPHP\test\integration;
 
-class EnvironmentService
+use heidelpayPHP\Traits\CanRecur;
+
+class DummyResource
 {
-    const ENV_VAR_NAME_ENVIRONMENT = 'HEIDELPAY_MGW_ENV';
-    const ENV_VAR_VALUE_STAGING_ENVIRONMENT = 'STG';
-    const ENV_VAR_VALUE_DEVELOPMENT_ENVIRONMENT = 'DEV';
-    const ENV_VAR_VALUE_PROD_ENVIRONMENT = 'PROD';
-
-    public function getMgwEnvironment()
-    {
-        return getenv(self::ENV_VAR_NAME_ENVIRONMENT);
-    }
+    use CanRecur;
 }