+++ test/unit/Traits/TraitDummyCanRecur.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * This service provides for functionalities concerning the mgw environment.
+ * This class defines a dummy implementing CanRecur trait.
  *
  * Copyright (C) 2019 heidelpay GmbH
  *
@@ -20,19 +20,19 @@
  *
  * @author  Simon Gabriel <development@heidelpay.com>
  *
- * @package  heidelpayPHP/services
+ * @package  heidelpayPHP/test/unit
  */
-namespace heidelpayPHP\Services;
+namespace heidelpayPHP\test\unit\Traits;
 
-class EnvironmentService
+use heidelpayPHP\Resources\AbstractHeidelpayResource;
+use heidelpayPHP\Traits\CanRecur;
+
+class TraitDummyCanRecur extends AbstractHeidelpayResource
 {
-    const ENV_VAR_NAME_ENVIRONMENT = 'HEIDELPAY_MGW_ENV';
-    const ENV_VAR_VALUE_STAGING_ENVIRONMENT = 'STG';
-    const ENV_VAR_VALUE_DEVELOPMENT_ENVIRONMENT = 'DEV';
-    const ENV_VAR_VALUE_PROD_ENVIRONMENT = 'PROD';
+    use CanRecur;
 
-    public function getMgwEnvironment()
+    public function getId()
     {
-        return getenv(self::ENV_VAR_NAME_ENVIRONMENT);
+        return 'myId';
     }
 }